+++ database/migrations/2022_06_02_144311_create_time_trackers_table.php
@@ -1,0 +1,38 @@
+<?php
+
+use App\Models\User;
+use App\Models\Contract;
+use Illuminate\Support\Facades\Schema;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class CreateTimeTrackersTable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::create('time_trackers', function (Blueprint $table) {
+            $table->id();
+            $table->foreignIdFor(User::class);
+            $table->foreignIdFor(Contract::class);
+            $table->string('task_title');
+            $table->dateTime('start');
+            $table->dateTime('end');
+            $table->timestamps();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::dropIfExists('time_trackers');
+    }
+}