+++ database/migrations/2022_06_06_113557_add_screenshot_duration_to_screenshots.php
@@ -1,0 +1,32 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class AddScreenshotDurationToScreenshots extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('screenshots', function (Blueprint $table) {
+            $table->integer('screenshot_duration');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('screenshots', function (Blueprint $table) {
+            $table->dropColumn('screenshot_duration');
+        });
+    }
+}