+++ resources/views/contents/freelancer/work-diary.blade.php
@@ -21,18 +21,29 @@
                 </select>
             </div>
             <div class="col-8"></div>
-            <div class="col-4 mt-4">
-                <h4>History for:</h4>
-                <input class="form-control" type="date" name="history_of" id="history_of" value="{{date('Y-m-d')}}">
-            </div>
-            <div class="col-12 mt-5">
-                <div class="card">
-                    <div class="card-body row" id="history_body">
-                        <div class="fa-3x">
-                            <i class="fa-solid fa-circle-notch fa-spin"></i>
-                        </div>
+            <div class="col-12 mt-4">
+                <div class="row justify-content-between">
+                    <div class="col-auto"">
+                        <h4>History for:</h4>
+                        <input class="form-control" type="date" name="history_of" id="history_of" value="{{date('Y-m-d')}}">
+                    </div>
+                    <div class="col-8 align-self-end text-end">
+                        <button class="btn btn-primary border-zero" id="delete_screenshot">Delete</button>
                     </div>
                 </div>
+            </div>
+            <div class="col-12 mt-5">
+                <form action="#" id="work_history_form">
+                    <div class="card">
+                        <div class="card-body row" id="history_body">
+                            <div>
+                                <div class="fa-3x">
+                                    <i class="fa-solid fa-circle-notch fa-spin"></i>
+                                </div>
+                            </div>
+                        </div>
+                    </div>
+                </form>
             </div>
         </div>
     @endif
@@ -45,5 +56,7 @@
         history_of.addEventListener('change', getWorkHistory);
 
         window.onload = getWorkHistory;
+        document.getElementById('delete_screenshot').addEventListener('click', deleteScreenshot);
+        
     </script>
 @endpush