+++ routes/web.php
@@ -296,6 +296,7 @@ Route::group(['middleware' => ['auth:web,admin,staff', 'user.activity']], functi
     Route::prefix('f/work-diary')->group(function () {
         Route::get('/', [FreelancerWorkDiaryController::class, 'index'])->name('freelancer.work.diary.index');
         Route::get('/{contract}/{date}', [FreelancerWorkDiaryController::class, 'show'])->name('freelancer.work.diary.show');
+        Route::DELETE('/', [FreelancerWorkDiaryController::class, 'destroy'])->name('freelancer.work.diary.destroy');
     });
     
     /**